Implemented the requested changes
