+++ src/main.rs
@@ -5,19 +5,16 @@
 
 use eframe::egui;
 
-mod icon;
-
 // When compiling natively:
 #[cfg(not(target_arch = "wasm32"))]
 fn main() {
-    
     let native_options = eframe::NativeOptions {
         always_on_top: true,
         decorated: true,
         initial_window_size: Some(egui::vec2(550f32, 309f32)),
         resizable: false,
         transparent: true,
-        icon_data: Some(icon::load_icon("assets/icon-64x64.ico")),
+        icon_data: Some(load_icon("assets/icon-64x64.ico")),
         ..Default::default()
     };
     eframe::run_native(
@@ -28,4 +25,20 @@ fn main() {
             Box::new(rusty_autoclicker::RustyAutoClickerApp::new(cc))
         }),
     );
+}
+pub fn load_icon(path: &str) -> eframe::IconData {
+    let (icon_rgba, icon_width, icon_height) = {
+        let image = image::open(path)
+            .expect("Failed to open icon path")
+            .into_rgba8();
+        let (width, height) = image.dimensions();
+        let rgba = image.into_raw();
+        (rgba, width, height)
+    };
+
+    eframe::IconData {
+        rgba: icon_rgba,
+        width: icon_width,
+        height: icon_height,
+    }
 }
